CSchoolSite/requirements.txt
django
django-tinymce
django-bootstrap3
django-registration
